requirements.txt
streamlit
cohere
pandas
numpy
matplotlib
requests
python-dotenv
Pillow
scikit-learn
librosa
openai-whisper
